// src/app/shared/models/productRules.model.ts
export class ProductRules {
  rulSeq: number;
  rulId: number;
  rulCtgryKey: number;
  rulNm: string;
  rulCmnt: string;
  rulCrtraStr: string;
  isAdded = false;

}
export class AdvanceRules {
  prdAdvRulSeq: number;
  rulSeq: number;
  prdSeq: number;
  rulId: number;
  rulCtgryKey: number;
  rulNm: string;
  rulCmnt: string;
  rulCrtraStr: string;
  isAdded = false;
}